Explicit conversions to short for GDALColorEntry.
There probably needs to be a check to make sure there are never negative
color values.  Added TODOs.


git-svn-id: https://svn.osgeo.org/gdal/trunk@36010 f0d54148-0727-0410-94bb-9a71ac55c965

+++ gdal/frmts/gtiff/geotiff.cpp
@@ -12780,11 +12780,12 @@ CPLErr GTiffDataset::OpenOffset( TIFF *hTIFFIn,
 
         for( int iColor = nColorCount - 1; iColor >= 0; iColor-- )
         {
+            // TODO(schwehr): Ensure the color entries are never negative?
             const short divisor = 256;
             const GDALColorEntry oEntry = {
-                panRed[iColor] / divisor,
-                panGreen[iColor] / divisor,
-                panBlue[iColor] / divisor,
+                static_cast<short>(panRed[iColor]) / divisor,
+                static_cast<short>(panGreen[iColor]) / divisor,
+                static_cast<short>(panBlue[iColor]) / divisor,
                 bNoDataSet &&
                 static_cast<int>(dfNoDataValue) == iColor
                 ? static_cast<short>(0)
@@ -12809,10 +12810,11 @@ CPLErr GTiffDataset::OpenOffset( TIFF *hTIFFIn,
 
             for( int iColor = nColorCount - 1; iColor >= 0; iColor-- )
             {
+                // TODO(schwehr): Ensure the color entries are never negative?
                 const GDALColorEntry oEntry = {
-                    panRed[iColor],
-                    panGreen[iColor],
-                    panBlue[iColor],
+                    static_cast<short>(panRed[iColor]),
+                    static_cast<short>(panGreen[iColor]),
+                    static_cast<short>(panBlue[iColor]),
                     bNoDataSet &&
                     static_cast<int>(dfNoDataValue) == iColor
                     ? static_cast<short>(0)
